feat(workflow): unified frontmatter+body parser
Replace the two format-specific parsers (parser.ts's # Workflow:/##
Step: grammar, program/parser.ts's YAML field validator) with one
frontmatter+body parser targeting the unified WorkflowDocument shape.

Frontmatter carries the orchestration graph (params/defaults/budget/
steps), adopting the program vocabulary nearly verbatim minus
version/name/title/instructions/gate.criteria, plus new inputs: on
unit/map steps. The body binder enforces the three rules: every H2
must be `## <step-id>` for a declared step; unit/map steps require a
body section, route steps may have one; an optional `### gate`
sub-heading inside a step's section is the rubric (the format's one
reserved marker) — frontmatter gate: without a rubric is a lint error,
a rubric alone declares a default gate.

Semantic checks retargeted at frontmatter: duplicate ids, route target
existence/forward-only/no-self-route/unique matches, timeout format,
retry-reason taxonomy, resource limits (validator.ts keeps only xrefs +
resource-limit checks now that the JSON Schema owns the closed key
set — ALLOWED_FRONTMATTER_KEYS/checkFrontmatterKeys are gone).

program/schema.ts sheds instructions/title/criteria/version/name and
gains ProgramStep.inputs; program/parser.ts and program/project.ts
(the YAML-program-only field validator and projections) are deleted.
